User Profile Function
1. show basic information
2. update info
3. set about me visibility (update sql schema)

+++ cbse-blog-server/src/main/java/com/loloao/service/UserService.java
@@ -22,7 +22,7 @@ public interface UserService extends IService<User> {
 
     Long saveUser(User user);
 
-    Long updateUser(User user);
+    int updateUser(User user);
 
     void deleteUserById(Long id);
 }
